samples: zephyr: Split out tests from test runner
Move the list of tests to run from the test runner itself into its own
package.  The test can be invoked the same way (Go will find the go.mod
at the top of mcuboot), but this will make it easier to separate a test
runner from a test builder.

Signed-off-by: David Brown <david.brown@linaro.org>