bootutil: add swap move with size config
Use config MCUBOOT_SWAP_MOVE_SIZE to give a size that should be used to
perform the move-up and swap operations. This way a size that is a
valid multiple of the sizes existing in both the primary and secondary
can be used to make them compatible, or to move larger blocks of data on
each operation.

Signed-off-by: Fabio Utzig <fabio.utzig@nordicsemi.no>