boot: add support for custom user code to handle bootloader events
An event callback is invoked at each stage of the bootloader
lifecycle to indicate progress, errors, and image swap operations.
The user application can provide the function `bootloader_event`
to receives these events and performs design specific tasks such
as updating UI indicators to show image swap progress or bootloader
error.

`USER_ADDON_FILES` can be used to pass a list of user source files,
with custom code to handle events, to cmake be built as part of the
bootloader.

This feature enables users to customize some aspects of the bootloader
for their own application and board without needing to modify
the bootloader source code and keeping it in sync the master repo.

Signed-off-by: Arvin Farahmand <arvinf@ip-logix.com>