Add reference counting to Mbed OS flash backend
Backported patch, hash of the original commit: d1233e1b2dd82e6b64dfcc539967bd40185b2a06

The internal functions of mcuboot are not consistent in opening/closing flash areas and sometimes nested calls happen to `flash_area_open` and `flash_area_close`. With the previous implementation, a nested call to `flash_area_close` would deinitialize the underlying `BlockDevice`. This could cause subsequent flash operations on an "open" flash area to fail.

This PR adds a simple open counter for each flash area and ensures the underlying `BlockDevice` is initialized and deinitialized appropriately. The `BlockDevice` is only initialized when transitioning from an open count of 0 to 1. The `BlockDevice` is only deinitialized when the open count falls to 0.

Change-Id: I52dacff8710e4f79ff535c0f14bd8f854587b058
Signed-off-by: George Beckstein <becksteing@embeddedplanet.com>