ptest: Use github workflow instead of travis
Now that the travis description is empty, ptest no longer runs any
tests.  Change it to use the github workflow description.  It is still
fairly specific to our particular workflow, and will need to be updated
if the form of what is there changes.

The workflow description is quite a bit simpler, so a bunch of code has
been removed.

Signed-off-by: David Brown <david.brown@linaro.org>