boot: zephyr: Switch to chosen zephyr,uart-mcumg for serial recovery
The commit changes the default device selection, for serial recovery,
from the DTS chosen zephyr,console to zephyr,uart-mcumgr.
This is more in line with Zephyr which uses the later for mcumgr
UART, which is also used for serial recovery in mcuboot.

The default configurations of Zephyr boards select the same device,
uart0, for both chosen devices; projects that do not use overlays to
change these selections will not be affected.

Signed-off-by: Dominik Ermel <dominik.ermel@nordicsemi.no>