Fix version comparison result causing boot failure
Backported patch, hash of the original commit: d4d90f8eea61cb38dc23f5d05fe06e5a0d5a86cf

The variable `rc` is used in `context_boot_go` for a number of operations, mostly checking return codes for an error. It is also used to store the result of comparing the installed application version numbers (using `boot_version_cmp`).

At the end of `context_boot_go`, `rc` is returned as the result of the operation.

In some configurations, namely direct XIP (without revert or any other extra boot checks), it is possible for the comparison result to be the last value of `rc`. In most cases, this will cause `context_boot_go` to report a failure (ie: non-zero return code), as the value of `rc` will be 1 or -1 if the installed applications aren't the same version.

This commit resets the value of `rc` to 0 after the version comparison has taken place.

Fixes #976

Change-Id: Ibca96a152dd30d71d38a21b538a8d7e6dc85eccb
Signed-off-by: George Beckstein <george.beckstein@gmail.com>