Add configuration parameter to enable internal MCUboot logging.
Prior to this commit, it was not possible for a user to enable `MCUBOOT_HAVE_LOGGING` through the existing Mbed-OS configuration parameters. This commit fixes this by adding an option: `mcuboot.enable-logging`. Now, logging that is performed by MCUboot's internal sources will be printed as well as logging that is part of the Mbed-OS port.

Signed-off-by: George Beckstein <george.beckstein@gmail.com>