Mbed-OS: Prevent use of BlockDevice with unreliable erase value
The default return value of `BlockDevice::get_erase_value` is -1. When implicitly cast from an `int` to a `uint8_t`, this gets truncated to `0xFF`, which is normally a reasonable erase value.

However, the intent of the Mbed-OS BlockDevice API is that an **integer** value of -1 is to be interpreted as an "unreliable" erase value. ie: the underlying BlockDevice does not have a known erase value.

If a user attempts to use such a BlockDevice (eg: SDBlockDevice at the moment, see ARMmbed/mbed-os#15070), they may encounter hard to debug errors.

Since MCUboot expects some fields to be the "erased value" under certain boot/swap states, a BlockDevice with an unreliable erase value is not compatible with MCUboot. It is recommended that the user wrap the incompatible BlockDevice with a FlashSimBlockDevice (available in Mbed's sources) that ensures a known erase value.

This commit adds an assert that will (hopefully) catch such an error during development.

Signed-off-by: George Beckstein <george.beckstein@gmail.com>