zephyr: Phase out UPDATEABLE_IMAGE_NUMBER as user selectable option
The UPDATEABLE_IMAGE_NUMBER is numeric value which makes it impossible
to be used as a condition in other Kconfig options.
This commit introduces new Kconfig options:
 SUPPORTED_UPDATABLE_IMAGES_ONE
 SUPPORTED_UPDATABLE_IMAGES_TWO
that correspond to supporting one or two images, for DFU, respectively;
these options can be easily used as conditions for applying other
Kconfig options.
The UPDATEABLE_IMAGE_NUMBER will be deprecated, although it will not
be removed; it will become non-selectable and its value will depend
on selection of the two new options.

Signed-off-by: Dominik Ermel <dominik.ermel@nordicsemi.no>