boot_serial: Sw downgrade prevention in serial recovery configuration
In a serial recovery configuration, software based downgrade prevention
option cannot be used for now as it is designed to trigger a software
version comparison only before a swap in a 2 slots MCUboot configuration

In fact, serial recovery update will overwrite 1st slot on every update,
 so this check won't be done on the reception of a new firmware.

This commit tackles this issue as it implements the comparison of
software versions on the reception of the first update chunk (which
contains the image header).

So, if CONFIG_MCUBOOT_DOWNGRADE_PREVENTION option is selected, a
comparison will be done between stored and incoming firmware.

If the stored firmware version is strictly superior to incoming one,
flash writing will be bypassed and an error code will be sent to upgrade
sender. Otherwise, the update process is done and stored firmware
is overwritten.

Signed-off-by: Mateo Ligne <mateo.ligne@netatmo.com>