samples/zephyr: Remove the body of the shell test runner
Instead of trying to main two test runners for the Zephyr sample
directory, have the old test runner just print out a message that users
should use the new test runner.  Eventually, we can just remove old
runner entirely.

Instead of asking the user for each test if it worked or not, the new
test runner looks at the serial log and makes this conclusion itself.
It is able to run the whole test suite completely automatically.

Signed-off-by: David Brown <david.brown@linaro.org>