mbed: Add the size of the MCU boot header area
Add header area size at the start of FLASH parameter in mbed_lib.json
This parameter is not used in MCUboot source code directly but for
example is the argument for imgtool.py sign command which sign
the binary image.
It will be useful to have it in Mbed port settings to use during project
building. For example post build command in cmake which signs image.

Signed-off-by: Artur Tynecki <artur.tynecki@mobica.com>