boot: zephyr: Force module name to `mcuboot`
The Zephyr build system exposes to module CMakeLists.txt files
the following variables:

  ZEPHYR_${MODULE_NAME}_MODULE_DIR
  ZEPHYR_${MODULE_NAME}_CMAKE_DIR

where MODULE_NAME is the uppercase version of the current module.
This variable is set based on the `name:` field in `module.yml`.
If the `name:` field is not present, then it picks up the name
from the manifest file if present, and follows other defaulting
logic if not.  In order to enable other modules to reference
overlays in this module the module name must be fixed by setting
it in `module.yml`.

Verified by:
1. setting the name to the value "mcuboot2" and verifying
   the build artifacts produced `$build/modules/mcuboot2`.
2. setting the name to the value "mcuboot" and verifying
   the build artifacts produced `$buildmodules/mcuboot`
3. Further verified that the following testcases still pass with
   mcuboot options on twister:
      mcuboot/boot/zephyr
      zephyr/tests/subsys/dfu
      zephyr/samples/subsys/mgmt/mcumgr/smp_svr

Signed-off-by: Gregory SHUE <gregory.shue@legrand.com>