boot: zephyr: boot/zephyr/include/ to zephyr/include/
In the Zephyr directory structure the module-level include is
located as a subdirectory of the module-level CMakeLists.txt.
For a general pattern suitable for extending existing repositories
the Zephyr directory structure must exist within the `zephyr/`
directory along with `module.yml`.  This is necessary to avoid
conflicts with existing repository directory structures and
build files.

Verified by:
1. Building `boot/zephyr` for mimxrt1064_evk and verifying
   that the zephyr image did not change in sizes.
2. Further verified that the following testcases still pass with
   mcuboot options on twister:
      mcuboot/boot/zephyr
      zephyr/tests/subsys/dfu
      zephyr/samples/subsys/mgmt/mcumgr/smp_svr

Signed-off-by: Gregory SHUE <gregory.shue@legrand.com>