New sample: mcuboot_external_config
Created a sample of how to configure MCUBoot build from an
external repository.  This sample provides no `main()` definition
and reuses all logic from MCUBoot and other repositories.  The
functionality matches the existing functionality of mcuboot/boot/zephyr.

The sample is located within `$mcuboot/zephyr/samples/modules/mcuboot/`
so that the directory path matches that used in Zephyr for samples
of external modules.  The directory name `mcuboot_external_config`
intentionally begins with the module name to help avoid futures
conflict with samples that may be defined in other repositories.
Together, this pattern supports easy integration of this sample
documentation with sample documentation in Zephyr.

Attempted Verification:

1. west build -p always -b nrf52840dk_nrf52840 \
            mcuboot/zephyr/samples/modules/mcuboot/mcuboot_external_config/

Signed-off-by: Gregory Shue <gregory.shue@legrand.com>