Refactor Kconfig into project/module parts
- Add prompt to MCUBOOT Kconfig and default 'n'
- Add prompt to MCUBOOT_APP Kconfig and default 'n'
- Add prompt to MCUBOOT_DEVICE_SETTINGS_ Kconfig and default 'n'.
  These device settings rules need to not be applied for the
  psa_crypto build.
- Add boot/zephyr/Kconfig reference to module.yml
- Relocate module Kconfig into mcuboot/zephyr/
- Relocate module-level Kconfig declarations into the module's
  zephyr/ subdirectory so that it is visible for external build
  configurations.  Removed sourcing of Kconfig.zephyr
- Restore sourcing of Kconfig.zephyr to stripped boot/zephyr/Kconfig.

The MCUBOOT Kconfig seems to control the inclusion of MCUboot
application code in the Zephyr build.. Because this name matches
the module and an alternate name was used to include MCUboot's
bootutil package the name usage is a bit unorthodox.

An ortodox and scalable pattern is for a Kconfig matching the
module name to control the exposure of the module itself
(e.g., the public headers).  The first step in this transition is
to make MCUBOOT public and default to 'n'.  Defaulting to 'n' will
prevent this module with interfering with other existing builds.

Incremental verification:

1. (Pass) west build -p always -b nrf52840dk_nrf52840 mcuboot/boot/zephyr/
2. (Pass) ./zephyr/scripts/twister --testsuite-root mcuboot/boot/zephyr/
3. (Pass) ./zephyr/scripts/twister --testsuite-root zephyr/samples/tfm_integration/psa_crypto/
4. (Pass) ./zephyr/scripts/twister --testsuite-root zephyr/tests/subsys/dfu/
5. (Pass) ./zephyr/scripts/twister --testsuite-root zephyr/samples/subsys/mgmt/mcumgr/smp_svr/

Signed-off-by: Xiang Xiao <xiaoxiang@xiaomi.com>
Signed-off-by: Gregory Shue <gregory.shue@legrand.com>